pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "siqspeak"
version = "0.1.0"
description = "Local speech-to-text for Windows — hold a hotkey, speak, release, auto-type"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "SIQstack"}]
keywords = ["speech-to-text", "whisper", "dictation", "windows"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Win32 (MS Windows)",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "faster-whisper",
    "sounddevice",
    "numpy",
    "pystray",
    "pillow",
    "pyperclip",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pyright>=1.1",
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
siqspeak = "siqspeak.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "UP",
    "B",
    "SIM",
    "RUF",
]
ignore = [
    "E501",
]

[tool.ruff.lint.isort]
known-first-party = ["siqspeak"]

[tool.ruff.format]
quote-style = "double"

[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "basic"
include = ["src"]
exclude = ["tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
